pyproject.toml

[tool.poetry]
name = "uh50"
version = "0.1.0"
description = "Read out UltraHeat 50"
authors = ["Berry <berrydenhartog@mailbox.org>"]
license = "Apache-2.0"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
Flask = "^2.1.2"
Flask-Caching = "^2.0.0"
pydantic = "^1.9.1"
dateutils = "^0.6.12"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"
coverage = {extras = ["toml"], version = "^6.4.1"}
pylint = "^2.14.4"
pytest-mock = "^3.8.1"
flake8 = "^4.0.1"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--strict-markers"

markers = [
    "slow"
]
testpaths = [
    "tests"
]

[tool.black]
target-version = ['py310']

[tool.coverage.paths]
source = ["uh50"]

[tool.coverage.run]
branch = true
source = ["uh50"]
omit = [
    "uh50\\ultraheat.py",
    "uh50\\__main__.py"
    ]

[tool.coverage.report]
fail_under = 99